+++ packages/plugin/disk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @umac-js/disk
 
+## 2.2.7
+
+### Patch Changes
+
+- Add browser default to open plugin
+
+- Updated dependencies []:
+  - @umac-js/utils@2.2.7
+
 ## 2.2.6
 
 ### Patch Changes